+++ src/main/kotlin/com/crypted/explorer/service/transaction/TransactionMongoRepository.kt
@@ -17,9 +17,9 @@ interface TransactionMongoRepository : MongoRepository<TransactionMongoDO, Strin
 
     fun findByToAndStatus(to: String, status: Int, pageable: Pageable): Page<TransactionMongoDO>
 
-    fun findByBlockNumber(blockNumber: Int, pageable: Pageable): Page<TransactionMongoDO>
+    fun findByBlockNumber(blockNumber: Long, pageable: Pageable): Page<TransactionMongoDO>
 
-    fun findByBlockNumberAndStatus(blockNumber: Int, status: Int, pageable: Pageable): Page<TransactionMongoDO>
+    fun findByBlockNumberAndStatus(blockNumber: Long, status: Int, pageable: Pageable): Page<TransactionMongoDO>
 
     fun findByHash(hash: String): TransactionMongoDO?
 
@@ -31,9 +31,9 @@ interface TransactionMongoRepository : MongoRepository<TransactionMongoDO, Strin
 
     fun countByToAndStatus(to: String, status: Int): Int
 
-    fun countByBlockNumberAndStatus(blockNumber: Int, status: Int): Int
+    fun countByBlockNumberAndStatus(blockNumber: Long, status: Int): Int
 
-    fun countByBlockNumber(blockNumber: Int): Int
+    fun countByBlockNumber(blockNumber: Long): Int
 
     fun countByStatus(status: Int): Int
